requirements.txt
appdirs==1.4.3
boto3==1.16.30
botocore==1.19.30
distlib==0.3.0
filelock==3.0.12
importlib-metadata==1.5.0
jmespath==0.10.0
json-rpc==1.13.0
more-itertools==4.2.0
PyMySQL==0.10.1
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.4
s3transfer==0.3.3
six==1.14.0
SQLAlchemy==1.3.20
urllib3==1.26.2
virtualenv==20.0.17
Werkzeug==1.0.1
zipp==1.0.0
requests==2.25.0
ptvsd